refactored java files into packages
