--- src/main/java/DukeException.java
@@ -1,9 +1,0 @@
-/**
- * Represents duke exceptions.
- */
-public class DukeException extends Exception{
-    public String description;
-    public DukeException (String description) {
-        this.description = description;
-    }
-}
