+++ src/main/java/command/CommandExit.java
@@ -1,3 +1,10 @@
+package command;
+
+import command.Command;
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
+
 /**
  * Represents the exit command to exit the program.
  */
