+++ src/main/java/exception/DukeException.java
@@ -1,0 +1,11 @@
+package exception;
+
+/**
+ * Represents duke exceptions.
+ */
+public class DukeException extends Exception{
+    public String description;
+    public DukeException (String description) {
+        this.description = description;
+    }
+}
